requirements.txt
streamlit
openai
faiss-cpu
pypdf
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
tiktoken